MPfm 0.5.2.1 - Finished adding and correcting all XML comments on all projects.
Related to issue #287.


git-svn-id: http://hamster-svn/svn/repos/base@300 765c1f7c-9fb8-954f-9ff8-dd0915cb3117

+++ MPfm/branches/current/MPfm.Sound/Tags/SV8Metadata.cs
@@ -244,7 +244,7 @@ namespace MPfm.Sound
             long value = 0;
             for (intLength = 1; intLength <= 9; intLength++)
             {
-                int currentByte = (int)data[intLength - 1];
+                long currentByte = (long)data[intLength - 1];
                 value = (value << 7);
                 value = (value | (currentByte & 0x7F));
                 if ((currentByte & 0x80) == 0)
@@ -267,7 +267,7 @@ namespace MPfm.Sound
             long value = 0;
             for (intLength = 1; intLength <= 9; intLength++)
             {
-                int currentByte = (int)reader.ReadByte();
+                long currentByte = (long)reader.ReadByte();
                 value = (value << 7);
                 value = (value | (currentByte & 0x7F));
                 if ((currentByte & 0x80) == 0)
